# Conflicts:
#   docker/box4security.dev.yml
#   docker/box4security.yml

+++ docker/openvas/scripts/insertconfig.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
-python3 -m venv .venv-openvas
-source .venv-openvas/bin/activate
-pip install python-gvm
-python /root/config.py
-deactivate
-echo "OpenVAS Config Full and Fast without Default Account Check and Bruteforce imported."
-rm -r .venv-openvas
+CONFIG=/etc/openvas/4s-OpenVAS.xml
+
+if [ -f "$FILE" ]; then
+  python3 -m venv .venv-openvas
+  source .venv-openvas/bin/activate
+  pip install python-gvm
+  python /root/config.py
+  deactivate
+  echo "OpenVAS Config Full and Fast without Default Account Check and Bruteforce imported."
+  rm -r .venv-openvas
+  rm CONFIG
+else
+  echo "Config has already been imported"
+fi
